+++ IntegrationBus/source/sim/eth/EthController.hpp
@@ -86,12 +86,10 @@ private:
         CallbackVector<EthTransmitAcknowledge>
     > _callbacks;
 
-
     PcapTracer _tracer;
     bool _tracingIsEnabled{false};
 
-     std::vector<std::pair<uint64_t, EthTxId>> _pendingAcks;
-
+    std::vector<std::pair<EthMac, EthTxId>> _pendingAcks;
 };
 
 // ================================================================================