+++ IntegrationBus/source/sim/eth/EthControllerTest.cpp
@@ -87,6 +87,7 @@ TEST_F(EthernetControllerTest, send_eth_message)
         .Times(1);
 
     EthMessage msg;
+    msg.ethFrame.SetSourceMac(EthMac{ 0,0,0,0,0,0 });
     controller.SendMessage(msg);
 }
 
@@ -96,6 +97,7 @@ TEST_F(EthernetControllerTest, send_eth_message)
 TEST_F(EthernetControllerTest, trigger_callback_on_receive_message)
 {
     EthMessage msg;
+    msg.ethFrame.SetSourceMac(EthMac{ 0,0,0,0,0,0 });
 
     EXPECT_CALL(callbacks, ReceiveMessage(&controller, msg))
         .Times(1);
@@ -103,16 +105,19 @@ TEST_F(EthernetControllerTest, trigger_callback_on_receive_message)
     controller.ReceiveIbMessage(otherAddress, msg);
 }
 
-/*! \brief Passing an Ack to an EthControllers must trigger the registered callback
+/*! \brief Passing an Ack to an EthControllers must trigger the registered callback, iff
+ *         it sent a message with corresponding transmit ID and source MAC
  */
 TEST_F(EthernetControllerTest, trigger_callback_on_receive_ack)
 {
-    EthTransmitAcknowledge ack{17, 0x0, 42ms, EthTransmitStatus::Transmitted};
+    EthMessage msg;
+    msg.ethFrame.SetSourceMac(EthMac{1,2,3,4,5,6});
+    auto tid = controller.SendMessage(msg);
 
+    EthTransmitAcknowledge ack{tid, EthMac{1,2,3,4,5,6}, 42ms, EthTransmitStatus::Transmitted};
     EXPECT_CALL(callbacks, MessageAck(&controller, ack))
         .Times(1);
-    EthMessage msg;
-    msg.ethFrame.SetSourceMac(EthMac{ 0,0,0,0,0,0 });
+
     controller.ReceiveIbMessage(otherAddress, ack);
 }
 
@@ -128,9 +133,9 @@ TEST_F(EthernetControllerTest, generate_ack_on_receive_msg)
     EthMessage msg;
     msg.timestamp  = 42ms;
     msg.transmitId = 17;
-    msg.ethFrame.SetSourceMac(EthMac{ 0,0,0,0,0,0 });
+    msg.ethFrame.SetSourceMac(EthMac{1,2,3,4,5,6});
 
-    EthTransmitAcknowledge expectedAck{msg.transmitId, 0x0, msg.timestamp, EthTransmitStatus::Transmitted};
+    EthTransmitAcknowledge expectedAck{ msg.transmitId, EthMac{1,2,3,4,5,6}, msg.timestamp, EthTransmitStatus::Transmitted };
 
     EXPECT_CALL(comAdapter, SendIbMessage(controllerAddress, expectedAck))
         .Times(1);